requirements.txt
Flask==2.1.3
line-bot-sdk==2.3.0
pandas==1.4.3
gspread==5.4.0
oauth2client==4.1.3
python-dotenv==0.20.0
